git-svn-id: svn://svn.code.sf.net/p/meddly/code/branches/holes@428 14966d26-f10a-4722-9547-d5dcc201c352

+++ src/operations/comp_eq.cc
@@ -47,7 +47,7 @@ class MEDDLY::equal_mdd : public generic_binary_mdd {
       expert_forest* arg2, expert_forest* res);
 
   protected:
-    virtual bool checkTerminals(int a, int b, int& c);
+    virtual bool checkTerminals(long a, long b, long& c);
 };
 
 MEDDLY::equal_mdd::equal_mdd(const binary_opname* opcode, 
@@ -57,7 +57,7 @@ MEDDLY::equal_mdd::equal_mdd(const binary_opname* opcode,
   operationCommutes();
 }
 
-bool MEDDLY::equal_mdd::checkTerminals(int a, int b, int& c)
+bool MEDDLY::equal_mdd::checkTerminals(long a, long b, long& c)
 {
   if (arg1F->isTerminalNode(a) &&
       arg2F->isTerminalNode(b)) {
@@ -86,7 +86,7 @@ class MEDDLY::equal_mxd : public generic_binbylevel_mxd {
       expert_forest* arg2, expert_forest* res);
 
   protected:
-    virtual bool checkTerminals(int a, int b, int& c);
+    virtual bool checkTerminals(long a, long b, long& c);
 };
 
 MEDDLY::equal_mxd::equal_mxd(const binary_opname* opcode, 
@@ -96,7 +96,7 @@ MEDDLY::equal_mxd::equal_mxd(const binary_opname* opcode,
   operationCommutes();
 }
 
-bool MEDDLY::equal_mxd::checkTerminals(int a, int b, int& c)
+bool MEDDLY::equal_mxd::checkTerminals(long a, long b, long& c)
 {
   if (arg1F->isTerminalNode(a) &&
       arg2F->isTerminalNode(b)) {
@@ -125,8 +125,8 @@ class MEDDLY::equal_evtimes : public generic_binary_evtimes {
       expert_forest* arg2, expert_forest* res);
 
   protected:
-    virtual bool checkTerminals(float av, int a, float bv, int b, 
-      float &cv, int& c);
+    virtual bool checkTerminals(float av, long a, float bv, long b, 
+      float &cv, long& c);
 };
 
 MEDDLY::equal_evtimes::equal_evtimes(const binary_opname* opcode, 
@@ -137,7 +137,7 @@ MEDDLY::equal_evtimes::equal_evtimes(const binary_opname* opcode,
 }
 
 bool MEDDLY::equal_evtimes
-::checkTerminals(float aev, int a, float bev, int b, float &cev, int& c)
+::checkTerminals(float aev, long a, float bev, long b, float &cev, long& c)
 {
   if (arg1F->isTerminalNode(a) &&
       arg2F->isTerminalNode(b)) {
